requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
beautifulsoup4==4.12.2
lxml==4.9.3
pyautogui==0.9.54
keyboard==0.13.5
psutil==5.9.6
schedule==1.2.0
colorama==0.4.6
